requirements.txt
aioice==0.10.1
aiortc==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
av==14.4.0
cffi==2.0.0
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.121.3
google-crc32c==1.7.1
h11==0.16.0
idna==3.11
ifaddr==0.2.0
jwcrypto==1.5.6
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
pylibsrtp==0.12.0
pymongo==4.15.4
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python-jwt==4.1.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
